Work on mean height map creator
I seem to actually be able to get the mean height of a pixel. Without crashing. More work to do when it comes to actually creating the height map, but this is an excellent starting point (unless I made a mistake and it doesn't actually work and I just think it does for some reason)